Broke out state machine parsing code into separate file.
